tests/wpt/web-platform-tests/css/css-grid/grid-model/grid-container-ignores-first-letter-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: '::first-letter' is ignored in grid containers</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="http://www.w3.org/TR/css-grid-1/#grid-containers">
<link rel="help" href="http://www.w3.org/TR/css3-selectors/#first-letter">
<meta name="assert" content="The test checks that grid item should ignore grid container's first-letter pseudo-element.">
<link href="../support/grid.css" rel="stylesheet">
<style>
  body { line-height: 20px; }
  .grid-first-letter::first-letter { line-height: 100px; color: red; }
  .container-first-letter::first-letter { line-height: 200px; color: green; }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.container');">

<div id="log"></div>

<div class="container">
  <div class="grid grid-first-letter">
    <div class="item" data-expected-height=20>The first item.</div>
    <div class="item" data-expected-height=20>The second item.</div>
  </div>
</div>

<div class="container">
  <div class="inline-grid grid-first-letter">
    <div class="item" data-expected-height=20>The first item.</div>
    <div class="item" data-expected-height=20>The second item.</div>
  </div>
</div>

<div class="container">
  <div class="grid grid-first-letter" data-expected-height=20>
    Anonymous item.
  </div>
</div>

<div class="container">
  <div class="inline-grid grid-first-letter" data-expected-height=20>
    Anonymous item.
  </div>
</div>

<div class="container container-first-letter">
  <div class="grid grid-first-letter">
    <div class="item" data-expected-height=20>The first item.</div>
    <div class="item" data-expected-height=20>The second item.</div>
  </div>
  <div data-expected-height=200>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="inline-grid grid-first-letter">
    <div class="item" data-expected-height=20>The first item.</div>
    <div class="item" data-expected-height=20>The second item.</div>
  </div>
  <div data-expected-height=20>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="grid grid-first-letter" data-expected-height=20>
    Anonymous item.
  </div>
  <div data-expected-height=200>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="inline-grid grid-first-letter" data-expected-height=20>
    Anonymous item.
  </div>
  <div data-expected-height=20>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="grid">
    <div class="item" data-expected-height=20>The first item.</div>
    <div class="item" data-expected-height=20>The second item.</div>
  </div>
  <div data-expected-height=200>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="inline-grid">
    <div class="item" data-expected-height=20>The first item.</div>
    <div class="item" data-expected-height=20>The second item.</div>
  </div>
  <div data-expected-height=20>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="grid" data-expected-height=20>
    Anonymous item.
  </div>
  <div data-expected-height=200>Out of grid.</div>
</div>

<div class="container container-first-letter">
  <div class="inline-grid" data-expected-height=20>
    Anonymous item.
  </div>
  <div data-expected-height=20>Out of grid.</div>
</div>

</body>
