LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-model/grid-container-ignores-first-letter-001-expected.txt


PASS .container 1
PASS .container 2
PASS .container 3
PASS .container 4
FAIL .container 5 assert_equals:
<div class="container container-first-letter">
  <div class="grid grid-first-letter">
    <div class="item" data-expected-height="20">The first item.</div>
    <div class="item" data-expected-height="20">The second item.</div>
  </div>
  <div data-expected-height="20">Out of grid.</div>
</div>
height expected 20 but got 200
PASS .container 6
FAIL .container 7 assert_equals:
<div class="container container-first-letter">
  <div class="grid grid-first-letter" data-expected-height="20">
    Anonymous item.
  </div>
  <div data-expected-height="20">Out of grid.</div>
</div>
height expected 20 but got 200
PASS .container 8
FAIL .container 9 assert_equals:
<div class="container container-first-letter">
  <div class="grid">
    <div class="item" data-expected-height="20">The first item.</div>
    <div class="item" data-expected-height="20">The second item.</div>
  </div>
  <div data-expected-height="20">Out of grid.</div>
</div>
height expected 20 but got 200
PASS .container 10
FAIL .container 11 assert_equals:
<div class="container container-first-letter">
  <div class="grid" data-expected-height="20">
    Anonymous item.
  </div>
  <div data-expected-height="20">Out of grid.</div>
</div>
height expected 20 but got 200
PASS .container 12
The first item.
The second item.
The first item.
The second item.
Anonymous item.
Anonymous item.
The first item.
The second item.
Out of grid.
The first item.
The second item.
Out of grid.
Anonymous item.
Out of grid.
Anonymous item.
Out of grid.
The first item.
The second item.
Out of grid.
The first item.
The second item.
Out of grid.
Anonymous item.
Out of grid.
Anonymous item.
Out of grid.
